adding of capitalize support of all commands, some comments and refactoring
