+++ requirements.txt
@@ -1,3 +1,3 @@
 python-telegram-bot==5.3.0
 redis==2.10.5
--e git://github.com/squeaky-pl/japronto#egg=japronto
+tornado
